sweet alert added for delete

+++ src/Pages/MyEquipment.jsx
@@ -1,6 +1,7 @@
-import React, { useContext, useState, useEffect } from 'react';
+import React, { useContext, useState } from 'react';
 import { useLoaderData } from 'react-router-dom';
 import { AuthContext } from '../Provider/AuthProvider';
+import Swal from 'sweetalert2';
 
 const MyEquipment = () => {
   const { user } = useContext(AuthContext);
@@ -10,16 +11,33 @@ const MyEquipment = () => {
   const filterEquipment = equipment.filter(equip => equip.email === user.email);
 
   const handleDelete = (_id) => {
-    fetch(`http://localhost:5000/addEquipment/${_id}`, {
-      method: 'DELETE'
-    })
-    .then(res => res.json())
-    .then(data => {
-      if(data.deletedCount > 0){
-        alert('hello')
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        fetch(`http://localhost:5000/addEquipment/${_id}`, {
+          method: 'DELETE'
+        })
+        .then(res => res.json())
+        .then(data => {
+          if (data.deletedCount > 0) {
+       
+            setEquipment(prevEquipment => prevEquipment.filter(equip => equip._id !== _id));
+            Swal.fire(
+              "Deleted!",
+              "Your equipment has been deleted.",
+              "success"
+            );
+          }
+        });
       }
-  
-      setEquipment(prevEquipment => prevEquipment.filter(equip => equip._id !== _id));
     });
   };
